Move test to entities_client_test.go
