+++ internal/onboard/service_sync_test.go
@@ -52,133 +52,6 @@ const testDTEntityQueryResponse = `{
     ]
 }`
 
-func Test_serviceSynchronizer_fetchKeptnManagedServicesFromDynatrace(t *testing.T) {
-
-	var returnedEntitiesResponse dynatrace.EntitiesResponse
-	dtMockServer := httptest.NewServer(http.HandlerFunc(func(writer http.ResponseWriter, request *http.Request) {
-		marshal, _ := json.Marshal(returnedEntitiesResponse)
-		writer.WriteHeader(200)
-		writer.Write(marshal)
-	}))
-
-	defer dtMockServer.Close()
-
-	type fields struct {
-		projectsAPI     *keptnapi.ProjectHandler
-		servicesAPI     *keptnapi.ServiceHandler
-		resourcesAPI    *keptnapi.ResourceHandler
-		apiMutex        sync.Mutex
-		EntitiesClient  *dynatrace.EntitiesClient
-		syncTimer       *time.Ticker
-		keptnHandler    *keptnv2.Keptn
-		servicesInKeptn []string
-	}
-	type args struct {
-		nextPageKey string
-		pageSize    int
-	}
-	tests := []struct {
-		name                         string
-		fields                       fields
-		args                         args
-		want                         []dynatrace.Entity
-		returnedEntitiesListResponse dynatrace.EntitiesResponse
-		wantErr                      bool
-	}{
-		{
-			name: "",
-			fields: fields{
-				projectsAPI:  nil,
-				servicesAPI:  nil,
-				resourcesAPI: nil,
-				apiMutex:     sync.Mutex{},
-				EntitiesClient: dynatrace.NewEntitiesClient(dynatrace.NewDynatraceHelper(nil, &credentials.DTCredentials{
-					Tenant:   dtMockServer.URL,
-					ApiToken: "",
-				})),
-				syncTimer:       nil,
-				keptnHandler:    nil,
-				servicesInKeptn: nil,
-			},
-			args: args{
-				nextPageKey: "",
-				pageSize:    1,
-			},
-			want: []dynatrace.Entity{
-				{
-					EntityID:    "1",
-					DisplayName: "name",
-					Tags: []dynatrace.Tag{
-						{
-							Context:              "CONTEXTLESS",
-							Key:                  "keptn_managed",
-							StringRepresentation: "keptn_managed",
-							Value:                "",
-						},
-						{
-							Context:              "CONTEXTLESS",
-							Key:                  "keptn_service",
-							StringRepresentation: "keptn_service:my-service",
-							Value:                "my-service",
-						},
-					},
-				},
-			},
-			returnedEntitiesListResponse: dynatrace.EntitiesResponse{
-				TotalCount:  1,
-				PageSize:    1,
-				NextPageKey: "",
-				Entities: []dynatrace.Entity{
-					{
-						EntityID:    "1",
-						DisplayName: "name",
-						Tags: []dynatrace.Tag{
-							{
-								Context:              "CONTEXTLESS",
-								Key:                  "keptn_managed",
-								StringRepresentation: "keptn_managed",
-								Value:                "",
-							},
-							{
-								Context:              "CONTEXTLESS",
-								Key:                  "keptn_service",
-								StringRepresentation: "keptn_service:my-service",
-								Value:                "my-service",
-							},
-						},
-					},
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s := &serviceSynchronizer{
-				projectsAPI:     tt.fields.projectsAPI,
-				servicesAPI:     tt.fields.servicesAPI,
-				resourcesAPI:    tt.fields.resourcesAPI,
-				EntitiesClient:  tt.fields.EntitiesClient,
-				syncTimer:       tt.fields.syncTimer,
-				keptnHandler:    tt.fields.keptnHandler,
-				servicesInKeptn: tt.fields.servicesInKeptn,
-			}
-			returnedEntitiesResponse = tt.returnedEntitiesListResponse
-			got, err := s.EntitiesClient.GetKeptnManagedServices()
-			if (err != nil) != tt.wantErr {
-				t.Errorf("fetchKeptnManagedServicesFromDynatrace() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if diff := deep.Equal(got, tt.want); len(diff) > 0 {
-				t.Errorf("fetchKeptnManagedServicesFromDynatrace() got = %v, want %v", got, tt.want)
-				for _, d := range diff {
-					t.Log(d)
-				}
-			}
-		})
-	}
-}
-
 func Test_doesServiceExist(t *testing.T) {
 	type args struct {
 		services    []string
